Split modules and plugins apart
