+++ lua/custom/options.lua
@@ -52,3 +52,5 @@ opt.signcolumn = 'yes'
 -- Enable break indent
 opt.breakindent = true
 
+-- The line beneath this is called `modeline`. See `:help modeline`
+-- vim: ts=2 sts=2 sw=2 et
